+++ src/main/java/org/cidd/dlaas/initialization/GlorotUniform.java
@@ -10,14 +10,14 @@ import org.nd4j.linalg.api.ndarray.INDArray;
 public class GlorotUniform implements Initializer {
 
     @Override
-    public INDArray handle(int[] size) {
-        Integer[] decomposed = decomposeSize(size);
+    public INDArray handle(long[] size) {
+        Long[] decomposed = decomposeSize(size);
         double scale = Math.sqrt(6.0d / (decomposed[0].doubleValue() + decomposed[1].doubleValue()));
         return Uniform.builder().scale(scale).build().handle(size);
     }
 
-    private Integer[] decomposeSize(int[] size) {
-        Integer[] decomposed = new Integer[2];
+    private Long[] decomposeSize(long[] size) {
+        Long[] decomposed = new Long[2];
         if (size.length == 2) {
             decomposed[0] = size[0];
             decomposed[1] = size[1];
@@ -28,7 +28,7 @@ public class GlorotUniform implements Initializer {
             decomposed[0] = respectiveFieldSize * size[0];
             decomposed[1] = respectiveFieldSize * size[1];
         } else {
-            int respectiveFieldSize = 0;
+            long respectiveFieldSize = 0;
             for (int i = 0; i < size.length; i++)
                 respectiveFieldSize *= size[i];
             respectiveFieldSize = (int)Math.sqrt(respectiveFieldSize);
