+++ src/main/java/org/cidd/dlaas/initialization/Initializer.java
@@ -5,6 +5,6 @@ import org.nd4j.linalg.api.ndarray.INDArray;
 
 public interface Initializer {
 
-    INDArray handle(int[] size);
+    INDArray handle(long[] size);
 
 }
